Solid
Hi Nick,

Your application is a good choice. You'll get more experience in all the areas we've covered this year (user interface design and database access). It's do-able and well described.

Small issues with spelling in a few places. See my note on dependencies.

Good work. This is a level 4 proposal.

Mr. Gordon